changes for export to float16

+++ QEfficient/transformers/models/glm4_moe/modeling_glm4_moe.py
@@ -133,7 +133,7 @@ def eager_attention_forward_blocked_kv(
 
     # Initialize Running Maximum
     batch_size, num_heads, seq_len, _ = query.shape
-    current_max = torch.full((batch_size, num_heads, seq_len), float(MIN_MASKED_ATTENTION_VALUE))
+    current_max = torch.full((batch_size, num_heads, seq_len), (MIN_MASKED_ATTENTION_VALUE).to(query.dtype))
 
     # Initialize Denominator
     current_denominator = torch.zeros(batch_size, num_heads, seq_len)
@@ -141,7 +141,7 @@ def eager_attention_forward_blocked_kv(
     past_seen_tokens = cache_kwargs.get("past_seen_tokens")
     position_ids = cache_kwargs.get("position_ids")
     block_size = -(-past_seen_tokens // num_kv_blocks)
-    masked_tensor = torch.tensor(MIN_MASKED_ATTENTION_VALUE, dtype=torch.float32)
+    masked_tensor = torch.tensor(MIN_MASKED_ATTENTION_VALUE, dtype=query.dtype)
 
     for j in range(num_kv_blocks):
         start_index = j * block_size
@@ -207,10 +207,10 @@ def eager_attention_forward(
     attn_weights = torch.matmul(query, key_states.transpose(2, 3)) * scaling
     if attention_mask is not None:
         attn_weights = torch.where(
-            attention_mask, torch.tensor(MIN_MASKED_ATTENTION_VALUE, dtype=torch.float32), attn_weights
+            attention_mask, torch.tensor(MIN_MASKED_ATTENTION_VALUE, dtype=key_states.dtype), attn_weights
         )
 
-    attn_weights = nn.functional.softmax(attn_weights, dim=-1, dtype=torch.float32).to(query.dtype)
+    attn_weights = nn.functional.softmax(attn_weights, dim=-1, dtype=key_states.dtype).to(query.dtype)
     attn_output = torch.matmul(attn_weights, value_states)
     attn_output = attn_output.transpose(1, 2).contiguous()
 
@@ -778,7 +778,7 @@ class QEffGlm4MoeForCausalLM(Glm4MoeForCausalLM):
         hidden_states = outputs.last_hidden_state
         logit_index = position_ids.to(torch.int32).argmax(1, keepdim=True)
         hidden_states = hidden_states[torch.arange(position_ids.shape[0]).view(-1, 1), logit_index]
-        logits = self.lm_head(hidden_states).float()
+        logits = self.lm_head(hidden_states).to(hidden_states.dtype)
 
         return CausalLMOutputWithPast(
             loss=None,
